refactor: consolidate geo-utils
